[refactor/core]: replaced framework-core into the src directory
