+++ package.json
@@ -1,19 +1,19 @@
 {
-  "name": "my-ts-framework",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^4.1.1",
-    "@types/node": "^20.1.4",
-    "prettier": "^2.8.8",
-    "sass": "^1.62.1",
-    "typescript": "^5.0.2",
-    "vite": "^4.3.2"
-  }
+    "name": "my-ts-framework",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+        "dev": "vite --host",
+        "build": "tsc && vite build",
+        "preview": "vite preview"
+    },
+    "devDependencies": {
+        "@trivago/prettier-plugin-sort-imports": "^4.1.1",
+        "@types/node": "^20.1.4",
+        "prettier": "^2.8.8",
+        "sass": "^1.62.1",
+        "typescript": "^5.0.2",
+        "vite": "^4.3.2"
+    }
 }
