+++ src/framework-core/app/app.interface.ts
