+++ src/framework-core/component/base-component.abstract.interface.ts
