+++ src/main.ts
@@ -3,8 +3,11 @@ import { routes } from '@router/router'
 
 import './app/styles/main.scss'
 
+
 const app = new App({
 	rootElement: '#app'
 })
 
-app.useRoutes(routes).run()
+app
+	.useRoutes(routes)
+	.run()
