+++ src/views/home/home.view.module.ts
@@ -1,10 +1,10 @@
-import { BaseComponent } from '@core/component/base-component.abstract';
+import { BaseComponent } from '@core/component/base-component.abstract'
 
 import html from './home.component.html?raw'
 
 export class HomeView extends BaseComponent {
 
-	 render(): HTMLElement {
+	render(): HTMLElement {
 		const element = this.compiler.compile(html, [])
 
 		return element
