+++ tsconfig.json
@@ -24,12 +24,12 @@
 
     "paths": {
       /* Framework core aliases */
-      "@core/app/*": [ "core/app/*" ],
-      "@core/compiler/*": [ "core/comiler/*" ],
-      "@core/router/*": [ "core/router/*" ],
-      "@core/services/*": [ "core/services/*" ],
-      "@core/querify/*": [ "core/querify/*" ],
-      "@core/component/*": [ "core/component/*" ],
+      "@core/app/*": [ "src/framework-core/app/*" ],
+      "@core/compiler/*": [ "src/framework-core/compiler/*" ],
+      "@core/router/*": [ "src/framework-core/router/*" ],
+      "@core/services/*": [ "src/framework-core/services/*" ],
+      "@core/querify/*": [ "src/framework-core/querify/*" ],
+      "@core/component/*": [ "src/framework-core/component/*" ],
 
       /* Application aliases */
       "@entities/*": [ "src/entities/*" ],
@@ -42,5 +42,5 @@
 
     },
   },
-  "include": [ "src", "core" ]
+  "include": [ "src"]
 }
