+++ vite.config.js
@@ -4,12 +4,12 @@ export default defineConfig({
     resolve: {
         alias: {
             /* Framework core aliases */
-            '@core/router': '/core/router',
-            '@core/compiler': '/core/compiler',
-            '@core/querify': '/core/querify',
-            '@core/services': '/core/services',
-            '@core/app': '/core/app',
-            '@core/component': '/core/component',
+            '@core/router': '/src/framework-core/router',
+            '@core/compiler': '/src/framework-core/compiler',
+            '@core/querify': '/src/framework-core/querify',
+            '@core/services': '/src/framework-core/services',
+            '@core/app': '/src/framework-core/app',
+            '@core/component': '/src/framework-core/component',
 
             /* Application aliases */
             '@entities': '/src/entities',
